builder/requirements.txt
ipython==5.3.0
pytest==3.0.7
pytest-cov==2.4.0
spotipy==2.4.4
vcrpy==1.10.5
